+++ app/src/main/AndroidManifest.xml
@@ -37,7 +37,7 @@
          android:name=".SearchScreen.SearchDashboardActivity"
          android:exported="false" />
         <activity
-            android:name=".AddressScreen.PickAddress"
+            android:name=".Activity.MainActivity"
             android:exported="false" />
         <activity
             android:name=".AddressScreen.AddAddress"
@@ -64,10 +64,13 @@
             android:name=".SignUp.SignUp_1"
             android:exported="false" />
        <activity
-            android:name=".CartScreen.Cart"
+            android:name=".PaymentScreen.Payment"
             android:exported="false" />
         <activity
-            android:name=".Activity.MainActivity"
+            android:name=".AddressScreen.PickAddress"
+            android:exported="false" />
+        <activity
+            android:name=".CartScreen.Cart"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
