+++ app/src/main/java/com/group2/adapter/VoucherAdapter.java
@@ -1,4 +1,4 @@
-package com.group2.pop4u_app.AddressScreen.adapter;
+package com.group2.adapter;
 
 import android.app.Activity;
 import android.content.Context;
@@ -8,35 +8,34 @@ import android.view.ViewGroup;
 import android.widget.BaseAdapter;
 import android.widget.TextView;
 
-import com.group2.pop4u_app.AddressScreen.model.Address;
 import com.group2.pop4u_app.R;
-
+import com.group2.model.ItemVoucher;
 
 import java.util.List;
 
-public class AddressAdapter extends BaseAdapter {
+public class VoucherAdapter extends BaseAdapter {
     Activity context;
     int item_list;
-    List<Address> addressList;
+    List<ItemVoucher> voucherList;
 
     //Constructor
 
 
-    public AddressAdapter(Activity context, int item_list, List<Address> addressList) {
+    public VoucherAdapter(Activity context, int item_list, List<ItemVoucher> voucherList) {
         this.context = context;
         this.item_list = item_list;
-        this.addressList = addressList;
+        this.voucherList = voucherList;
     }
 
 
     @Override
     public int getCount() {
-        return addressList.size();
+        return voucherList.size();
     }
 
     @Override
     public Object getItem(int i) {
-        return addressList.get(i);
+        return voucherList.get(i);
     }
 
     @Override
@@ -46,33 +45,30 @@ public class AddressAdapter extends BaseAdapter {
 
     @Override
     public View getView(int i, View view, ViewGroup viewGroup) {
-        AddressAdapter.ViewHolder holder;
+        ViewHolder holder;
         if(view ==null) {
-            holder = new AddressAdapter.ViewHolder();
+            holder = new ViewHolder();
             LayoutInflater inflater = (LayoutInflater) context.getSystemService(Context.LAYOUT_INFLATER_SERVICE);
             view = inflater.inflate(item_list, null);
 
 
-            holder.cus_name = view.findViewById(R.id.cus_name);
-            holder.cus_phone = view.findViewById(R.id.cus_phone);
-            holder.cus_address = view.findViewById(R.id.cus_address);
+            holder.voucher_id = view.findViewById(R.id.voucher_id);
+            holder.voucher_description = view.findViewById(R.id.voucher_description);
             view.setTag(holder);
         }else{
-            holder = (AddressAdapter.ViewHolder)  view.getTag();
+            holder = (ViewHolder)  view.getTag();
 
         }
         //Lien ket du lieu
-        Address b = addressList.get(i);
-        holder.cus_name.setText(b.getCus_name());
-        holder.cus_phone.setText(b.getCus_phone());
-        holder.cus_address.setText(b.getCus_address());
+        ItemVoucher b = voucherList.get(i);
+        holder.voucher_id.setText(b.getVoucher_id());
+        holder.voucher_description.setText(b.getVoucher_description());
 
         return view;
     }
     public static class  ViewHolder{
-        TextView cus_name, cus_phone, cus_address;
+        TextView voucher_id, voucher_description;
     }
 }
-
