+++ local.properties
@@ -5,5 +5,5 @@
 # For customization when using a Version Control System, please read the
 # header note.
 #Wed Mar 27 00:14:07 ICT 2024
-sdk.dir=C\:\\Users\\Admin\\AppData\\Local\\Android\\Sdk
+sdk.dir=C\:\\Users\\Khanh Ha\\AppData\\Local\\Android\\Sdk
